More complete model classes.
Contact keeps now list of possible notifications using mobile phone numbers and
email addresses from system contacts database.  A notification may be enabled
or disable. Enabled notifications are used to send sms and email notifications.

Contact and Notification classes can be the data source for views. When
checking a notification checkbox, the notification should be enabled, and when
closing a view, dirty notifications should be updated in the database.