Implement contact validation
Contact information is copied from system contacts database to the application
database, to ensure that removing or modifying contact phones or emails will
invalidate contact notfications.

Each time we load contacts from the database, we remove phones or emails that
were removed from the system contacts database and add new phones and emails.
We assume that system contacts database is always up to date.

If enabled notifications were removed, disabling email or sms notifications for
a contact, we enable the first notifiction.