Set Contact and Notification id when adding to the database.
Previously Contact or Notification were not valid after adding to the database,
and could not be update in the database, becuase they had invalid id. It works
currently becuase we throw away the objects and create new instances by reading
from the database. This change will allow reading of all contacts from
the database once, and writing changes to the database when needed.