Cache contacts list and validate contacts only when needed.
Previously we read the contact list form the database and validated with the
system contacts database each time we display the contact list or after adding
a new contact or deleting existing contact.

Now the contact list is read once from the databse on the first access, and
never read again. Contacts are validated now with the system contacts database
before we display the contact list in the main activity, and when we enter
emergency mode.